build: remove FHIRServer connector (queue and lambda)
Ref. metriport/metriport-internal#2417
