--- packages/api/src/external/fhir/connector/connector.ts
@@ -1,24 +1,0 @@
-export type FHIRServerRequest = {
-  cxId: string;
-  patientId: string;
-  documentId: string;
-  payload: string;
-  requestId?: string;
-};
-
-export interface FHIRServerConnector {
-  upsertBatch(req: FHIRServerRequest): Promise<void>;
-}
-
-// TODO try to make `requestId` required
-export function createJobId(requestId: string | undefined, documentId: string): string {
-  return `${requestId}_${documentId}`;
-}
-
-export function parseJobId(
-  jobId?: string
-): { requestId?: string; documentId?: string } | undefined {
-  if (!jobId) return undefined;
-  const [requestId, documentId] = jobId.split("_");
-  return { requestId, documentId };
-}
