+++ packages/infra/lib/api-stack/api-service.ts
@@ -87,7 +87,6 @@ export function createAPIService({
   alarmAction,
   dnsZones,
   fhirServerUrl,
-  fhirServerQueueUrl,
   fhirConverterQueueUrl,
   fhirConverterServiceUrl,
   cdaToVisualizationLambda,
@@ -120,7 +119,6 @@ export function createAPIService({
   alarmAction: SnsAction | undefined;
   dnsZones: DnsZones;
   fhirServerUrl: string;
-  fhirServerQueueUrl: string | undefined;
   fhirConverterQueueUrl: string | undefined;
   fhirConverterServiceUrl: string | undefined;
   cdaToVisualizationLambda: ILambda;
@@ -260,9 +258,6 @@ export function createAPIService({
             FHIR_TO_CDA_CONVERTER_LAMBDA_NAME: fhirToCdaConverterLambda.functionName,
           }),
           FHIR_SERVER_URL: fhirServerUrl,
-          ...(fhirServerQueueUrl && {
-            FHIR_SERVER_QUEUE_URL: fhirServerQueueUrl,
-          }),
           ...(fhirConverterQueueUrl && {
             FHIR_CONVERTER_QUEUE_URL: fhirConverterQueueUrl,
           }),
