+++ packages/lambdas/src/sqs-to-converter.ts
@@ -4,7 +4,6 @@ import { DOC_ID_EXTENSION_URL } from "@metriport/core/external/fhir/shared/exten
 import { FHIR_APP_MIME_TYPE, TXT_MIME_TYPE, XML_APP_MIME_TYPE } from "@metriport/core/util/mime";
 import { errorToString, executeWithNetworkRetries, MetriportError } from "@metriport/shared";
 import { SQSEvent, SQSRecord } from "aws-lambda";
-import AWS from "aws-sdk";
 import axios from "axios";
 import * as uuid from "uuid";
 import { capture } from "./shared/capture";
@@ -12,7 +11,6 @@ import { CloudWatchUtils, Metrics } from "./shared/cloudwatch";
 import { getEnvOrFail } from "./shared/env";
 import { Log, prefixedLog } from "./shared/log";
 import { apiClient } from "./shared/oss-api";
-import { SQSUtils } from "./shared/sqs";
 import { cleanUpPayload } from "./sqs-to-converter/cleanup";
 
 // Keep this as early on the file as possible
@@ -25,7 +23,6 @@ const region = getEnvOrFail("AWS_REGION");
 const metricsNamespace = getEnvOrFail("METRICS_NAMESPACE");
 const apiURL = getEnvOrFail("API_URL");
 const axiosTimeoutSeconds = Number(getEnvOrFail("AXIOS_TIMEOUT_SECONDS"));
-const fhirServerQueueURL = getEnvOrFail("FHIR_SERVER_QUEUE_URL");
 const conversionResultBucketName = getEnvOrFail("CONVERSION_RESULT_BUCKET_NAME");
 
 const defaultS3RetriesConfig = {
@@ -33,7 +30,6 @@ const defaultS3RetriesConfig = {
   initialDelay: 500,
 };
 
-const sqs = new AWS.SQS({ region });
 const s3Utils = new S3Utils(region);
 const cloudWatchUtils = new CloudWatchUtils(region, lambdaName, metricsNamespace);
 const fhirConverter = axios.create({
@@ -158,7 +154,6 @@ export async function handler(event: SQSEvent) {
       const cxId = attrib.cxId?.stringValue;
       const patientId = attrib.patientId?.stringValue;
       const jobId = attrib.jobId?.stringValue;
-      const jobStartedAt = attrib.startedAt?.stringValue;
       const medicalDataSource = attrib.source?.stringValue;
       if (!cxId) throw new Error(`Missing cxId`);
       if (!patientId) throw new Error(`Missing patientId`);
@@ -278,7 +273,6 @@ export async function handler(event: SQSEvent) {
           patientId,
           s3FileName,
           updatedConversionResult,
-          jobStartedAt,
           jobId,
           medicalDataSource,
           log
@@ -359,7 +353,6 @@ async function sendConversionResult(
   patientId: string,
   sourceFileName: string,
   conversionPayload: FHIRBundle,
-  jobStartedAt: string | undefined,
   jobId: string | undefined,
   medicalDataSource: string | undefined,
   log: Log
@@ -382,26 +375,7 @@ async function sendConversionResult(
     }
   );
 
-  log(`Sending result info to queues`);
-  const queuePayload = JSON.stringify({
-    s3BucketName: conversionResultBucketName,
-    s3FileName: fileName,
-  });
-
-  await sqs
-    .sendMessage({
-      MessageBody: queuePayload,
-      QueueUrl: fhirServerQueueURL,
-      MessageAttributes: {
-        ...SQSUtils.singleAttributeToSend("cxId", cxId),
-        ...SQSUtils.singleAttributeToSend("patientId", patientId),
-        ...(jobStartedAt ? SQSUtils.singleAttributeToSend("jobStartedAt", jobStartedAt) : {}),
-        ...(jobId ? SQSUtils.singleAttributeToSend("jobId", jobId) : {}),
-        ...(medicalDataSource ? SQSUtils.singleAttributeToSend("source", medicalDataSource) : {}),
-      },
-    })
-    .promise();
-
+  log(`Sending result info to the API`);
   await ossApi.internal.notifyApi(
     { cxId, patientId, jobId, source: medicalDataSource, status: "success" },
     log
